Item Pickups & Drops, Game Over events, other tweaks
=== ITEMS & PICKUPS ===
- Implemented items / pickups using ItemPickupScript with CollisionScripts
- Item pickups have 2 colliders: Body and BodyHitbox
> Body is used as an "interactable" body, the body of all interactables in the world that can only interact with other interactable bodies (to prevent stacking on the same spot). Uses "InteractableBody" layer no.15
> BodyHitbox instead uses a new "Interactable" layer no.14 instead. This body hitbox acts as a hitbox that interacts with ActorBodies
NOTE: Layer no.12, formerly "Interactable" has been renamed to "Area" to represent some kind of background area of effect collider
- ItemPickupScript has an IItemPickup component; IItemPickup interface represents many different cases behaviour scripts for item pickups
> For example, Ammo pickup and Medkit pickup have their own IItemPickup behaviour scripts to make each item have different effects upon pickup

=== GAME OVER EVENTS ===
- FinishTrigger now invokes LevelManager new TryEndLevel method upon having both players and escortees passed it
- TryEndLevel method in LevelManager checks for all win conditions that have been met first before calling GameOver event in GameStateManager
- Added a new enum GameOverEvent, to determine which Game Over Event is executed. Consists of PERMADEATH, MISSION_SUCCESS, MISSION_FAIL, and ENDING

= MISC =
- Added isFinalMission flag in MissionData for PERMADEATH condition checking
- Added a new CompositeCollider2D for Blockade Groups
> This new CompositeCollider2D will be used by LevelManager to Update the A* Pathfinding Graphs after spawning blockades in the level/map
- Changed "amount" field in Ammo to be a property. Custom setter is set to clamp the amount between 0 and that ammo type's max amount
- Tweaked Spawnable weights from int to float

TODO:
- Move PERMADEATH and ENDING checks somewhere else. Maybe it's better to call for either event after a mission end instead of immediately calling the events right when a mission ends

+++ Assets/Scripts/Core/Simple Behaviours/Spawnable.cs
@@ -7,9 +7,9 @@ using System;
 public class Spawnable
 {
     public GameObject prefab;
-    [Range(1, 100)] public int spawnWeight;
+    [Range(1, 100)] public float spawnWeight;
 
-    public Spawnable(GameObject prefab, int spawnWeight)
+    public Spawnable(GameObject prefab, float spawnWeight)
     {
         this.prefab = prefab;
         this.spawnWeight = spawnWeight;
